List Integration
A rather huge update from a couple weeks of work maybe a month or so even.

I both added in a lot, but most recently is a system to actually allow users to go through lists, sublists, and vote for each translation.

The voting is what I am working on right now. This requires a small deck of functions in order to accomplish everything needed, but once that happens, we can have users properly vote and add novel translations to the database.