Visual Update - Themes and Data Viewer
Continued implementing Light Mode.
Changed Dark Mode to better co-exist and mirror Light Mode.
Updated Data viewer to show values and key.
Updated Data viewer to use colorblind-friendly colors in Light and Dark.
Updated Data viewer to accept options object with showDetails and isLarge - options can be extendend.
Finally, I am getting some bugs with the backend that I am trying to figure out, because while no errors are being returned, it seems that the backend functionality is not working as it is meant to.