Major updates
Many many updates, just realized I hadn't pushed. Mainly: new structure of data so that we can store each type of data in a table in supabase, access that table and store it as a tree structure, and then navigate these trees with an AddressBook with forms and sections so that given an address, we can get data from our trees.