Added Presets, backend and frontend
Basic presets added and set up on the backend and front.
Now, from /home/ users can choose a preset which changes something in their profile and refreshes the page.
Profiles store saved preset so that users come back always to the form they are interested in, but can switch it to whatever they want.